pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yt-bulk-transcripts"
version = "0.1.0"
description = "Production-grade bulk YouTube transcript collector (public captions only)"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "yt-dlp>=2024.0",
    "google-api-python-client>=2.100",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.4",
]

[project.scripts]
yt-bulk-transcripts = "yt_bulk_transcripts.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
